refactor(parser): extract grammar patterns and AST builders
Split parser.rs (710 LOC) into focused modules for better maintainability:

- parser/grammar.rs: Declarative grammar pattern constants and regex rules
- parser/builder.rs: AST node construction logic from matched patterns
- parser.rs: Core pattern matching engine (reduced to 307 LOC)

This separation improves code visibility and follows the same architectural
pattern established in the extraction module refactoring.

All 253 library tests pass.
